refactor: add hooks and components groups into 'wallet' feature
